split burgemeester-benoeming in multiple files
